# validator/nym/directory/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a temporarily centralized directory/PKI/metrics API to allow us to get the other Nym node types running. Its functionality will eventually be folded into other parts of Nym.",
        "title": "Nym Directory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/nymtech/nym-directory/license"
        },
        "version": "0.9.0-dev"
    },
    "paths": {
...
        "/api/presence/allow": {
            "post": {
                "description": "Sometimes when a node isn't working we need to temporarily remove it. This allows us to re-enable it once it's working again.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Removes a disallowed node from the disallowed nodes list",
                "operationId": "allow",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MixNodeID"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/presence/coconodes": {
            "post": {
                "description": "Nym Coconut nodes can ping this method to let the directory server know they're up. We can then use this info to create topologies of the overall Nym network.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Lets a coconut node tell the directory server it's alive",
                "operationId": "addCocoNode",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CocoHostInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/presence/disallow": {
            "post": {
                "description": "Sometimes when a node isn't working we need to temporarily remove it from use so that it doesn't mess up QoS for the whole network.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Takes a node out of the regular topology and puts it in the disallowed nodes list",
                "operationId": "disallow",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MixNodeID"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/presence/disallowed": {
            "get": {
                "description": "Sometimes we need to take mixnodes out of the network for repair. This shows which ones are currently disallowed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Lists Nym mixnodes that are currently disallowed",
                "operationId": "disallowed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MixNodePresence"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/presence/gateways": {
            "post": {
                "description": "Nym mix gateways can ping this method to let the directory server know they're up. We can then use this info to create topologies of the overall Nym network.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Lets a gateway tell the directory server it's alive",
                "operationId": "addGateway",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GatewayHostInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/presence/mixnodes": {
            "post": {
                "description": "Nym mixnodes can ping this method to let the directory server know they're up. We can then use this info to create topologies of the overall Nym network.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Lets mixnode a node tell the directory server it's alive",
                "operationId": "addMixNode",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MixHostInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/presence/topology": {
            "get": {
                "description": "Nym nodes periodically ping the directory server to register that they're alive. This method provides a list of nodes which have been most recently seen.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "Lists which Nym mixnodes, providers, gateways, and coconodes are alive",
                "operationId": "topology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Topology"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CocoHostInfo": {
            "type": "object",
            "required": [
                "pubKey",
                "type",
                "version"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.CocoPresence": {
            "type": "object",
            "required": [
                "lastSeen",
                "pubKey",
                "type",
                "version"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.GatewayHostInfo": {
            "type": "object",
            "required": [
                "clientListener",
                "identityKey",
                "mixnetListener",
                "sphinxKey",
                "version"
            ],
            "properties": {
                "clientListener": {
                    "type": "string"
                },
                "identityKey": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "mixnetListener": {
                    "type": "string"
                },
                "sphinxKey": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.GatewayPresence": {
            "type": "object",
            "required": [
                "clientListener",
                "identityKey",
                "lastSeen",
                "mixnetListener",
                "sphinxKey",
                "version"
            ],
            "properties": {
                "clientListener": {
                    "type": "string"
                },
                "identityKey": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "mixnetListener": {
                    "type": "string"
                },
                "sphinxKey": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.MixHostInfo": {
            "type": "object",
            "required": [
                "layer",
                "pubKey",
                "version"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "layer": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "models.MixNodeID": {
            "type": "object",
            "properties": {
                "pubKey": {
                    "type": "string"
                }
            }
        },
        "models.MixNodePresence": {
            "type": "object",
            "required": [
                "lastSeen",
                "layer",
                "pubKey",
                "version"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "lastSeen": {
                    "description": "MixStatusReport",
                    "type": "integer"
                },
                "layer": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.MixProviderPresence": {
            "type": "object",
            "required": [
                "lastSeen",
                "pubKey",
                "version"
            ],
            "properties": {
                "clientListener": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "mixnetListener": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "models.Topology": {
            "type": "object",
            "properties": {
                "cocoNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CocoPresence"
                    }
                },
                "gatewayNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GatewayPresence"
                    }
                },
                "mixNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MixNodePresence"
                    }
                },
                "mixProviderNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MixProviderPresence"
                    }
                }
            }
        }
    }
}